Relative imports interally. Moving calc_spot_price to analysis
