+++ lib/chainsync/chainsync/analysis/__init__.py
@@ -1,10 +1,5 @@
 """Analysis for trading."""
 from .calc_fixed_rate import calc_fixed_rate
 from .calc_ohlcv import calc_ohlcv
-from .calc_pnl import (
-    calc_closeout_pnl,
-    calc_single_closeout,
-    calc_total_returns,
-    calculate_spot_price,
-    calculate_spot_price_for_position,
-)
+from .calc_pnl import calc_closeout_pnl, calc_single_closeout, calc_total_returns
+from .calc_spot_price import calc_spot_price, calculate_spot_price_for_position
