+++ lib/chainsync/chainsync/dashboard/__init__.py
@@ -2,7 +2,7 @@
 
 from .build_leaderboard import build_leaderboard
 from .build_ticker import build_ticker
-from .extract_data_logs import calculate_spot_price, get_combined_data, read_json_to_pd
+from .extract_data_logs import get_combined_data, read_json_to_pd
 from .plot_fixed_rate import plot_fixed_rate
 from .plot_ohlcv import plot_ohlcv
 from .usernames import address_to_username, combine_usernames, get_user_lookup
