+++ lib/chainsync/chainsync/dashboard/build_ticker.py
@@ -1,6 +1,7 @@
 """Builds the ticker for the dashboard."""
 import pandas as pd
-from chainsync.dashboard import address_to_username
+
+from . import address_to_username
 
 
 def build_ticker(
