+++ lib/chainsync/chainsync/dashboard/extract_data_logs.py
@@ -24,11 +24,6 @@ def read_json_to_pd(json_file):
     return pd.DataFrame(json_data)
 
 
-def calculate_spot_price(share_reserves, bond_reserves, initial_share_price, time_stretch):
-    """Calculate the spot price."""
-    return ((initial_share_price * share_reserves) / bond_reserves) ** time_stretch
-
-
 def get_combined_data(txn_data, pool_info_data):
     """Combine multiple datasets into one containing transaction data, and pool info."""
     pool_info_data.index = pool_info_data.index.astype(int)
